+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/Consume.kt
@@ -1,8 +1,8 @@
-package world.gregs.voidps.world.activity.skill.cooking
+package world.gregs.voidps.world.activity.combat.consume
 
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.event.Event
 
-data class Consumable(val item: Item) : Event {
+data class Consume(val item: Item, val slot: Int) : Event {
     var cancel = false
 }
