+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/Drinking.kts
@@ -1,10 +1,11 @@
+package world.gregs.voidps.world.activity.combat.consume
+
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.entity.character.contain.inventory
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.hasEffect
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
 
 on<Consume>({ !cancel && (item.id.endsWith("_4") || item.id.endsWith("_3") || item.id.endsWith("_2") || item.id.endsWith("_1")) }, Priority.LOWER) { player: Player ->
     val doses = item.id.last().digitToInt()
