+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/Eating.kts
@@ -1,3 +1,5 @@
+package world.gregs.voidps.world.activity.combat.consume
+
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.entity.character.contain.inventory
 import world.gregs.voidps.engine.entity.character.player.Player
@@ -7,8 +9,6 @@ import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
 import world.gregs.voidps.engine.entity.hasOrStart
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consumable
-import world.gregs.voidps.world.activity.skill.cooking.Consume
 import world.gregs.voidps.world.interact.entity.player.equip.ContainerOption
 import world.gregs.voidps.world.interact.entity.sound.playSound
