+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/drink/Ale.kts
@@ -2,7 +2,7 @@ import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.entity.start
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
 on<Consume>({ item.id == "bandits_brew" }) { player: Player ->
     player.levels.boost(Skill.Thieving, 1)
