+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/drink/Antifire.kts
@@ -4,7 +4,7 @@ import world.gregs.voidps.engine.delay
 import world.gregs.voidps.engine.entity.*
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
 on<Registered> { player: Player ->
     player.restart("fire_resistance")
