+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/drink/Bottled.kts
@@ -1,7 +1,7 @@
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
 on<Consume>({ item.id == "karamjan_rum" }) { player: Player ->
     player.levels.boost(Skill.Strength, 5)
