+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/food/FishPie.kts
@@ -1,8 +1,8 @@
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
-on<Consume>({ item.id.startsWith("garden_pie") }) { player: Player ->
-    player.levels.boost(Skill.Farming, 5)
+on<Consume>({ item.id.startsWith("fish_pie") }) { player: Player ->
+    player.levels.boost(Skill.Fishing, 3)
 }
