+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/food/HolyBiscuits.kts
@@ -1,8 +1,8 @@
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
-on<Consume>({ item.id.startsWith("admiral_pie") }) { player: Player ->
-    player.levels.boost(Skill.Fishing, 5)
+on<Consume>({ item.id == "holy_biscuits" }) { player: Player ->
+    player.levels.restore(Skill.Prayer, 10)
 }
