+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/food/JugOfWine.kts
@@ -1,8 +1,8 @@
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
-on<Consume>({ item.id == "holy_biscuits" }) { player: Player ->
-    player.levels.restore(Skill.Prayer, 10)
+on<Consume>({ item.id == "jug_of_wine" }) { player: Player ->
+    player.levels.drain(Skill.Attack, 2)
 }
