+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/food/WildPie.kts
@@ -1,7 +1,7 @@
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.activity.skill.cooking.Consume
+import world.gregs.voidps.world.activity.combat.consume.Consume
 
 on<Consume>({ item.id.startsWith("wild_pie") }) { player: Player ->
     player.levels.boost(Skill.Slayer, 4)
